requirements.txt
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
bleach==3.1.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==4.0.2
django-cors-headers==3.11.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
drf-spectacular==0.21.2
drf-yasg==1.20.0
html5lib==1.1
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
kombu==5.2.3
MarkupSafe==2.1.0
packaging==21.3
phonenumbers==8.12.44
prompt-toolkit==3.0.28
pyparsing==3.0.7
PyPDF2==1.26.0
pyrsistent==0.18.1
python-bidi==0.4.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
reportlab==3.5.23
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.5
zipp==3.7.0
